requirements.txt
absl-py==2.2.2
accelerate==0.23.0
addict==2.4.0
ai2thor==4.0.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs @ file:///home/conda/feedstock_root/build_artifacts/appdirs_1733753955715/work
asciitree==0.3.3
asttokens==3.0.0
attrs==25.3.0
av==14.3.0
aws-requests-auth==0.4.3
basicsr==1.4.2
beautifulsoup4==4.13.4
blinker==1.9.0
botocore==1.38.1
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1648883617327/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1746569525376/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1636046055389/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1735929714516/work
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.0a10
dash==3.0.4
dds-cloudapi-sdk==0.4.1
debugpy @ file:///croot/debugpy_1736267418885/work
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
diffusers==0.34.0
dill==0.4.0
distro==1.9.0
docker-pycreds==0.4.0
einops==0.7.0
einops-exts==0.0.4
ema-pytorch==0.2.3
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1733327148154/work
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
facexlib==0.3.0
fastapi==0.115.12
fasteners==0.19
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1741969488311/work
filterpy==1.4.5
Flask==3.0.3
-e git+ssh://git@github.com/PlantPotatoOnMoon/Joint-Imagination-Environment-Feedback.git@21299df73c172dcbeb4e77b523df7e937baf197f#egg=flowdiffusion&subdirectory=flowdiffusion
fonttools==4.57.0
fsspec==2025.3.2
future==1.0.0
-e git+https://github.com/IDEA-Research/Grounding-DINO-1.5-API.git@2be80fe64ab3ca7f105821c31840ef303fbc14b9#egg=gdino
gdown==5.2.0
gfpgan==1.3.8
gitdb==4.0.12
GitPython==3.1.44
glfw==2.9.0
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1641732907770/work
gradio==4.22.0
gradio_client==0.13.0
grpcio==1.71.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
hydra-core==1.3.2
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython==8.18.1
ipywidgets==8.1.6
itsdangerous==2.2.0
jedi==0.19.2
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
jupyterlab_widgets==3.0.14
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.7.3
lpips==0.1.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1648737556467/work
matplotlib==3.8.0
matplotlib-inline==0.1.7
mdurl==0.1.2
-e git+ssh://git@github.com/PlantPotatoOnMoon/Joint-Imagination-Environment-Feedback.git@21299df73c172dcbeb4e77b523df7e937baf197f#egg=metaworld&subdirectory=metaworld
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
msgpack==1.1.0
mujoco-py==2.1.2.14
narwhals==1.36.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1698504735452/work
numba==0.60.0
numcodecs==0.12.1
numpy==1.24.4
nvidia-ml-py==12.570.86
nvitop==1.5.0
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1734769783178/work
omegaconf==2.3.0
open3d==0.19.0
openai==1.76.0
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.16
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.7
plotly==6.0.1
progressbar2==4.5.0
prompt_toolkit==3.0.51
protobuf==3.20.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.6.3
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.19.1
pynvml==11.5.0
pyparsing==3.2.3
pyquaternion==0.9.9
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-utils==3.9.1
python-xlib==0.33
pytorch-fid==0.3.0
pytz==2025.2
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1648757097602/work
pyzmq @ file:///croot/pyzmq_1705605076900/work
qwen-vl-utils==0.0.8
r3m @ git+https://github.com/facebookresearch/r3m@b2334e726887fa0206962d7984c69c5fb09cceab
realesrgan==0.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
retrying==1.3.4
rich==14.0.0
rpds-py==0.24.0
ruff==0.11.7
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.27.0
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.2
supervision==0.25.1
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1736248176451/work
tb-nightly==2.21.0a20250725
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==1.0.15
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.12.0
torch==2.2.0
torchaudio==2.2.0
torchvideotransforms @ git+https://github.com/hassony2/torch_videovision@785a810d1df2ccab6901fb98b38e358433e7b17a
torchvision==0.17.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827245914/work
tqdm==4.66.1
traitlets==5.14.3
transformers==4.51.3
triton==2.2.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1744302253/work
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.2
wandb==0.19.11
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.6
widgetsnbextension==4.0.14
yapf==0.43.0
zarr==2.18.2
zipp==3.21.0
zstandard @ file:///croot/zstandard_1677013143055/work
